% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCGPLVM.R
\name{fit.A}
\alias{fit.A}
\title{Fit function coefficients for target latent variable values.}
\usage{
fit.A(target, K.bc)
}
\arguments{
\item{K.bc}{}
}
\value{

}
\description{
Fit function coefficients for target latent variable values.
}

