% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCGPLVM.R
\name{fit.bcgplvm.sequential}
\alias{fit.bcgplvm.sequential}
\title{Fit a backconstrained GPLVM model sequentially}
\usage{
fit.bcgplvm.sequential(X, q, iterations = 1000, plot.freq = 100,
  classes = NULL, num.init.params = 100, K.bc.l = 0.1,
  Z.normal.prior = FALSE)
}
\arguments{
\item{Z.normal.prior}{}
}
\description{
Fit a backconstrained GPLVM model sequentially
}

