% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM.R
\name{fit.gplvm}
\alias{fit.gplvm}
\title{Fit an unconstrained GPLVM model}
\usage{
fit.gplvm(X, q, iterations = 1000, plot.freq = 100, classes = NULL,
  Z.init = NULL, num.init.params = 100, Z.normal.prior = FALSE)
}
\arguments{
\item{Z.normal.prior}{}
}
\description{
Fit an unconstrained GPLVM model
}

