% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticData.R
\name{generate.structured.dataset}
\alias{generate.structured.dataset}
\title{Generate structured synthetic data}
\usage{
generate.structured.dataset(num.samples = NULL, nrow, ncol,
  num.latent.variables = NULL, num.structures, function.noise = 0,
  nonlinear = FALSE, Z = NULL)
}
\arguments{
\item{Z}{}
}
\description{
Generate structured synthetic data
}

