% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisyComparison.R
\name{noisyComparison}
\alias{noisyComparison}
\title{Compare BCGPLVM to BCSGPLVM}
\usage{
noisyComparison(base.data, base.data.nrows, true.latent.vars, classes,
  noise.variances, PCA.init = FALSE, maxit = 1000)
}
\arguments{
\item{noise.variances}{}
}
\description{
Compare BCGPLVM to BCSGPLVM
}
