% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSA-BCSGPLVM.R
\name{predict.LSA_BCSGPLVM}
\alias{predict.LSA_BCSGPLVM}
\title{Predict data for an LSA_BCSGPLVM model}
\usage{
\method{predict}{LSA_BCSGPLVM}(object, data, training.data = NULL)
}
\arguments{
\item{training.data}{}
}
\description{
Predict data for an LSA_BCSGPLVM model
}
