% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSA-BCSGPLVM.R
\name{replay.plots}
\alias{replay.plots}
\title{Replay the plots from fitting an LSA_BCSGPLVM model}
\usage{
replay.plots(fit.lsa_bcsgplvm.output, time = 30, pps = 5)
}
\arguments{
\item{fit.lsa_bcsgplvm.output}{optimization to replay}

\item{time}{runtime}

\item{pps}{plots per second}
}
\description{
Replay the plots from fitting an LSA_BCSGPLVM model
}

