% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticData.R
\name{sample.from.model}
\alias{sample.from.model}
\title{Sample from a structured GPLVM model}
\usage{
sample.from.model(Z, nrow.X, K_S, K_Z)
}
\arguments{
\item{K_Z}{}
}
\description{
Sample from a structured GPLVM model
}
\examples{
Z <- rnorm(50, 2)
Z <- c(Z, rnorm(50, -2))
Z <- cbind(Z, rnorm(100))
K_Z <- GPLVM:::gplvm.SE(Z, 1, 1, 0)
K_S <- GPLVM:::structured.kernel.Matrix(50, 50, c(10,10), 1)
K_S <- K_S - Diagonal(nrow(K_S))
X <- sample.from.model(Z, 50, K_S, K_Z)
}
