% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCGPLVM.R
\name{select.bc.l.centile}
\alias{select.bc.l.centile}
\title{Select a lengthscale for constrained optimization by specifying the n'th
centile of the off-diagonal entries of the correlation matrix}
\usage{
select.bc.l.centile(X, params = NULL)
}
\arguments{
\item{X}{}

\item{params}{a vector with named entries "centile" and "target"}
}
\description{
Select a lengthscale for constrained optimization by specifying the n'th
centile of the off-diagonal entries of the correlation matrix
}
