% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCGPLVM.R
\name{select.bc.l.median}
\alias{select.bc.l.median}
\title{Select a lengthscale for constrained optimization by median}
\usage{
select.bc.l.median(X, target.median.cor)
}
\arguments{
\item{target.median.cor}{}
}
\description{
Select a lengthscale for constrained optimization by median
}
