% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCGPLVM.R
\name{select.bc.l.minmax}
\alias{select.bc.l.minmax}
\title{Select a lengthscale for constrained optimization by min and max}
\usage{
select.bc.l.minmax(X, target.max.min, target.min.max)
}
\arguments{
\item{target.min.max}{}
}
\description{
Select a lengthscale for constrained optimization by min and max
}

