% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM.R
\name{sparse.gplvm}
\alias{sparse.gplvm}
\title{Sparse GPLVM}
\usage{
sparse.gplvm(X, q, active.set.size, iterations = 10, Z.init = "PCA",
  classes = 1, Z.normal.prior = T)
}
\arguments{
\item{Z.normal.prior}{}
}
\description{
Sparse GPLVM
}
