% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuredGPLVM.R
\name{structured.gplvm.fixedC}
\alias{structured.gplvm.fixedC}
\title{Fit a structured GPLVM model with fixed structure covariance}
\usage{
structured.gplvm.fixedC(X, nrows.X, se.par = c(3, 1, 1),
  structured.par = c(20, 1, 1), Z = NULL, q = 2,
  probabilistic.trace.estimate = TRUE, maxit = 100, Z.normal.prior = TRUE)
}
\arguments{
\item{Z.normal.prior}{}
}
\description{
Fit a structured GPLVM model with fixed structure covariance
}
\examples{
Z.true <- cbind(rnorm(50, mean=-1), rnorm(50, mean=-1), rnorm(50, mean=-1))
Z.true <- rbind(Z.true, cbind(rnorm(50, mean=1), rnorm(50, mean=1), rnorm(50, mean=1)))
classes <- rep(1:2, each=50)
pairs(Z.true, col=classes, pty=".")
X <- generate.structured.dataset(100,20,20,num.latent.variables=NULL,5,0.1,nonlinear=TRUE, Z=Z.true)
X$data <- scale(X$data, center=T, scale=F)
X.sd <- sd(as.numeric(X$data))
X$data <- X$data / X.sd
X.noisy <- X$data + rnorm(length(X$data))
Z.init <- prcomp(X.noisy)$x[,1:5]
sgplvm.model <- structured.gplvm.fixedC(X=X$data, nrows.X=20, probabilistic.trace.estimate=F, structured.par=c(2, 2, 1), Z=Z.init, maxit=1000, q=5)
}
