% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{structured.kernel.Matrix}
\alias{structured.kernel.Matrix}
\title{Structured data kernel}
\usage{
structured.kernel.Matrix(nrows, ncols, C, alpha, grad = 0)
}
\arguments{
\item{grad}{}
}
\value{

}
\description{
Structured data kernel
}

